backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.6
reportlab==4.1.0
Pillow==10.2.0
httpx==0.26.0
python-dateutil==2.8.2
openpyxl==3.1.2
pytest==8.0.0
gunicorn==21.2.0
